file-organizer-web/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
alembic==1.13.1

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Scheduling
apscheduler==3.10.4

# File handling
python-magic==0.4.27
pillow==10.2.0
PyPDF2==3.0.1

# Date/Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Utilities
aiofiles==23.2.1
